requirements.txt
# Core framework
Django==5.1.6
asgiref==3.8.1

# REST API
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0

# Database & caching
psycopg2-binary==2.9.10
dj-database-url==2.3.0
django-redis==5.4.0
django-cacheops==7.1

# Django extensions & utilities
django-extensions==3.2.3
django-filter==25.1
django-cors-headers==4.7.0
django-js-asset==3.1.2
django-mptt==0.16.0

# Server & deployment
gunicorn==23.0.0
whitenoise==6.9.0

# HTTP clients
requests==2.32.3
httpx==0.27.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
idna==3.10
urllib3==2.3.0

# Parsing & scraping
beautifulsoup4==4.13.3
lxml==4.8.0
excelrd==3.0.0

# Schema & validation
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
fastjsonschema==2.21.1

# CLI & misc utilities
click==8.1.8
async-timeout==5.0.1
attrs==25.3.0
traitlets==5.14.3
typing_extensions==4.12.2
path==16.16.0
pathvalidate==3.2.3
platformdirs==4.3.6
six==1.17.0

# Jupyter support
jupyter_core==5.7.2
nbformat==5.10.4

# Logging & messaging
loguru==0.7.3
msgfy==0.2.1

# Markdown & frontmatter
Markdown==3.7
python-frontmatter==1.1.0

# Date/Time & config
python-dateutil==2.9.0.post0
pytz==2025.1
python-dotenv==1.0.1
envinfopy==0.2.0

# YAML
PyYAML==6.0.2

# Redis
redis==5.2.1

# Other utilities
DataProperty==1.1.0
funcy==2.0
packaging==24.2
protobuf==3.12.4
pytablereader==0.31.4
referencing==0.36.2
retryrequests==0.2.0
rpds-py==0.23.1
SimpleSQLite==1.5.4
soupsieve==2.6
sqlitebiter==0.36.3
sqliteschema==2.0.1
tabledata==1.3.4
tcolorpy==0.1.7
typepy==1.3.4
pillow-11.2.1


# Your .docx import libraries
python-docx>=1.1.2
mammoth>=1.9.0
